:sparkles: Derive dnote-mode from markdown-mode
Notes in dnote are markdown format, so inherit lots of goodness from
markdown-mode. Note though that I overwrite lots of the key bindings, on
purpose. I want to use C-c C-c to save a note. Markdown mode uses C-c C-c as
a prefix for lots of other commands. However, for my purposes anyway, all of
those commands aren't helpful when editing a dnote note.

So... let's go with this approach for now. It's fine for my needs. This
might change if other people ever use this and have an issue with it.